# Optimize checking gates